+++ example-es6.js
@@ -9,12 +9,12 @@ const logError = (message, ...metadata) => {
   console.error(chalk.bgBlack.red.bold(`\t\t ${message}`), ...metadata);
 };
 
-const deviceId = appgrid.generateUuid();
 const downloadsDirectoryName = 'downloads';
 
 const client = appgrid({
   appKey: '56ea6a370db1bf032c9df5cb',
-  deviceId,
+  // typically you wouldn't set the deviceId here, but let the default persistency strategy get it for you (with Express or a browser)
+  deviceId: 'fixedDeviceIDFromSDKSample',
   // log(...args) { console.log(...args); },
 });